+++ sdk/openai/azure-ai-openai/src/main/java/com/azure/ai/openai/models/EmbeddingItem.java
@@ -5,9 +5,14 @@ package com.azure.ai.openai.models;
 
 import com.azure.core.annotation.Generated;
 import com.azure.core.annotation.Immutable;
+import com.azure.core.util.BinaryData;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.nio.FloatBuffer;
+import java.util.ArrayList;
+import java.util.Base64;
 import java.util.List;
 
 /**
@@ -17,23 +22,29 @@ import java.util.List;
 public final class EmbeddingItem {
 
     /*
-     * List of embeddings value for the input prompt in base64 encoded string format. These represent
-     * a measurement of the vector-based relatedness of the provided input.
+     * List of embeddings value for the input prompt. These represent a measurement of the
+     * vector-based relatedness of the provided input.
      */
-    @Generated
     @JsonProperty(value = "embedding")
-    private List<Float> embedding;
+    private BinaryData embedding;
 
     /**
-     * Get the embedding property: List of embeddings value for the input prompt in base64 encoded string format. These
-     * represent
-     * a measurement of the vector-based relatedness of the provided input.
+     * Get the embedding property: List of embeddings value for the input prompt. These represent a measurement of the
+     * vector-based relatedness of the provided input.
      *
      * @return the embedding value.
      */
-    @Generated
     public List<Float> getEmbedding() {
-        return this.embedding;
+        return convertToFloatList(embedding.toString());
+    }
+
+    /**
+     * Get the embedding property: List of embeddings value in base64 format for the input prompt.
+     *
+     * @return the embedding base64 encoded string.
+     */
+    public String getEmbeddingString() {
+        return embedding.toString();
     }
 
     /*
@@ -60,9 +71,23 @@ public final class EmbeddingItem {
      * @param promptIndex the promptIndex value to set.
      */
     @JsonCreator
-    private EmbeddingItem(@JsonProperty(value = "embedding") List<Float> embedding,
+    private EmbeddingItem(@JsonProperty(value = "embedding") BinaryData embedding,
         @JsonProperty(value = "index") int promptIndex) {
         this.embedding = embedding;
         this.promptIndex = promptIndex;
+    }
+
+    private List<Float> convertToFloatList(String embedding) {
+        byte[] bytes = Base64.getDecoder().decode(embedding);
+        // View the raw binary data as floats (IEEE 754 binary32).
+        ByteBuffer byteBuffer = ByteBuffer.wrap(bytes);
+        byteBuffer.order(ByteOrder.LITTLE_ENDIAN);
+        FloatBuffer floatBuffer = byteBuffer.asFloatBuffer();
+        // Copy the floats in the buffer to a List representing an embedding.
+        List<Float> flaotList = new ArrayList<>(floatBuffer.remaining());
+        while (floatBuffer.hasRemaining()) {
+            flaotList.add(floatBuffer.get());
+        }
+        return flaotList;
     }
 }