+++ sdk/openai/azure-ai-openai/src/test/java/com/azure/ai/openai/NonAzureOpenAIAsyncClientTest.java
@@ -245,28 +245,6 @@ public class NonAzureOpenAIAsyncClientTest extends OpenAIClientTestBase {
 
     @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
     @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
-    public void getEmbeddingsWithFloatFormat(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
-        client = getNonAzureOpenAIAsyncClient(httpClient);
-        getEmbeddingsWithFloatFormatRunner((deploymentId, embeddingsOptions) -> {
-            StepVerifier.create(client.getEmbeddings(deploymentId, embeddingsOptions))
-                    .assertNext(resultEmbeddings -> assertEmbeddings(resultEmbeddings))
-                    .verifyComplete();
-        });
-    }
-
-    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
-    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
-    public void getEmbeddingsWithBase64Format(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
-        client = getNonAzureOpenAIAsyncClient(httpClient);
-        getEmbeddingsWithBase64FormatRunner((deploymentId, embeddingsOptions) -> {
-            StepVerifier.create(client.getEmbeddings(deploymentId, embeddingsOptions))
-                    .assertNext(resultEmbeddings -> assertEmbeddings(resultEmbeddings))
-                    .verifyComplete();
-        });
-    }
-
-    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
-    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
     public void testGetEmbeddingsWithResponse(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
         client = getNonAzureOpenAIAsyncClient(httpClient);
         getEmbeddingRunnerForNonAzure((modelId, embeddingsOptions) -> {